refactor: CustomLink to shared
