+++ src/app/components/index.ts
@@ -2,7 +2,6 @@ export { Actions } from './actions';
 export { CourseCard } from './course-card';
 export { DateLang } from './date-lang';
 export { EventCard, type EventCardProps } from './event-card';
-export { LinkCustom } from './link-custom';
 export { LogoWrapper } from './logo-wrapper';
 export { MobileView } from './mobile-view';
 export { OptionItem, type OptionItemProps } from './option-item';
