+++ src/features/course-main/course-main.tsx
@@ -1,11 +1,12 @@
 import { useLoaderData } from 'react-router-dom';
 import { getCourseStatus } from './utils';
-import { DateLang, LinkCustom, SectionLabel, Subtitle, Title } from '@/app/components';
+import { DateLang, SectionLabel, Subtitle, Title } from '@/app/components';
 import { useTitle } from '@/app/hooks';
 import { selectCourse } from '@/app/hooks/use-course-by-title/utils/select-course.ts';
 import { Course } from '@/app/types';
 import Image from '@/features/image';
 import { ArrowIcon } from '@/icons';
+import { LinkCustom } from '@/shared/ui/link-custom';
 
 import styles from './course-main.module.scss';
