+++ src/features/school/courses/courses.tsx
@@ -2,11 +2,11 @@ import { getCourseIcon } from './lib/getCourseIcon';
 import { IconsTitle } from './lib/icons.data';
 import { selectCourses } from './lib/selectCourses';
 import { CourseCard } from './ui/CourseCard';
-import { LinkCustom } from '@/app/components';
 import { ROUTES } from '@/app/const';
 import { useWindowSize } from '@/app/hooks';
 import { courses } from '@/app/services/data';
 import { ArrowIcon, RsBanner } from '@/icons';
+import { LinkCustom } from '@/shared/ui/link-custom';
 
 import './courses.scss';
